+++ sys/arch/powerpc/include/ansi.h
@@ -51,7 +51,7 @@
 #define	_BSD_PTRDIFF_T_		int		/* ptr1 - ptr2 */
 #define	_BSD_SIZE_T_		unsigned int	/* sizeof() */
 #define	_BSD_SSIZE_T_		int		/* byte count or error */
-#define	_BSD_TIME_T_		long		/* time() */
+#define	_BSD_TIME_T_		long long int	/* time() */
 
 /*
  * Actually va_list is an array, but this wouldn't work as a #define.
